Fixes compliance issue for NameMatchesFileName
