+++ Lusas_Engine/Compute/AssignmentWarning.cs
@@ -20,39 +20,36 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System.Collections.Generic;
-using System.Linq;
-using BH.oM.Structure.Elements;
-using BH.oM.Structure.Loads;
 using Lusas.LPI;
 
 namespace BH.Engine.Lusas
 {
-    public static partial class Convert
+    public partial class Compute
     {
-        public static BarTemperatureLoad ToBarTemperatureLoad(
-            IFLoading lusasTemperatureLoad,
-            IEnumerable<IFAssignment> lusasAssignments,
-            Dictionary<string, Bar> bars)
+        /***************************************************/
+        /**** Public methods                            ****/
+        /***************************************************/
+
+        public static void AssignmentWarning(IFAssignment lusasAssignment)
         {
-            IFLoadcase assignedLoadcase = (IFLoadcase)lusasAssignments.First().getAssignmentLoadset();
-            Loadcase bhomLoadcase = ToLoadcase(assignedLoadcase);
-            double temperatureChange = lusasTemperatureLoad.getValue("T")
-                - lusasTemperatureLoad.getValue("T0");
-
-            IEnumerable<Bar> bhomBars = Lusas.Query.GetLineAssignments(lusasAssignments, bars);
-            BarTemperatureLoad bhomBarTemperatureLoad = Structure.Create.BarTemperatureLoad(
-                bhomLoadcase,
-                temperatureChange,
-                bhomBars,
-                LoadAxis.Local,
-                false,
-                Lusas.Query.GetName(lusasTemperatureLoad));
-
-            int adapterID = Lusas.Query.GetAdapterID(lusasTemperatureLoad, 'l');
-            bhomBarTemperatureLoad.CustomData["Lusas_id"] = adapterID;
-            return bhomBarTemperatureLoad;
+            if (lusasAssignment.getDatabaseObject() is IFPoint)
+            {
+                Engine.Reflection.Compute.RecordWarning(
+                    lusasAssignment.GetType().ToString() + " does not support assignment to points, these have not been pulled");
+            }
+            else if (lusasAssignment.getDatabaseObject() is IFLine)
+            {
+                Engine.Reflection.Compute.RecordWarning(
+                    lusasAssignment.GetType().ToString() + "  does not support assignment to lines, these have not been pulled");
+            }
+            else if (lusasAssignment.getDatabaseObject() is IFSurface)
+            {
+                Engine.Reflection.Compute.RecordWarning(
+                    lusasAssignment.GetType().ToString() + " does not support assignment to surfaces, these have not been pulled");
+            }
         }
+
+        /***************************************************/
+
     }
 }
-
