+++ Lusas_Engine/Lusas_Engine.csproj
@@ -107,10 +107,15 @@
     <Compile Include="Create\MeshSettings1D.cs" />
     <Compile Include="Query\CheckIllegalCharacters.cs" />
     <Compile Include="Query\CheckPresets.cs" />
+    <Compile Include="Query\GetBar.cs" />
     <Compile Include="Query\GetBarRelease.cs" />
     <Compile Include="Query\GetConstraints.cs" />
-    <Compile Include="Query\GetDistinct.cs" />
+    <Compile Include="Query\GetDistinctPoints.cs" />
+    <Compile Include="Query\GetDistinctEdges.cs" />
+    <Compile Include="Query\GetEdge.cs" />
     <Compile Include="Query\GetFEAType.cs" />
+    <Compile Include="Query\GetGeometryAssignment.cs" />
+    <Compile Include="Query\GetLineAssignments.cs" />
     <Compile Include="Query\GetLoadAssignments.cs" />
     <Compile Include="Query\GetLoadcaseIDs.cs" />
     <Compile Include="Query\GetMeshProperties.cs" />
@@ -124,13 +129,14 @@
     <Compile Include="Convert\ToBHoM\Loads\ToBarDistributedLoad.cs" />
     <Compile Include="Query\GetAttributeAssignments.cs" />
     <Compile Include="Query\GetAdapterID.cs" />
-    <Compile Include="Query\GetAssignments.cs" />
-    <Compile Include="Query\GetLOF.cs" />
     <Compile Include="Query\GetName.cs" />
+    <Compile Include="Query\GetNode.cs" />
+    <Compile Include="Query\GetPointAssignments.cs" />
     <Compile Include="Query\GetObjectIDs.cs" />
+    <Compile Include="Query\GetSurfaceAssignments.cs" />
     <Compile Include="Query\IsMemberOf.cs" />
     <Compile Include="Modify\RemovePrefix.cs" />
-    <Compile Include="Compute\Warnings.cs" />
+    <Compile Include="Compute\AssignmentWarning.cs" />
     <Compile Include="Convert\ToBHoM\Properties\ToMeshSettings2D.cs" />
     <Compile Include="Convert\ToBHoM\Properties\ToMeshSettings1D.cs" />
     <Compile Include="Convert\ToBHoM\Loads\ToPointDisplacement.cs" />
