+++ Lusas_Engine/Query/GetEdge.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -21,51 +21,27 @@
  */
 
 using System.Collections.Generic;
-using System.Linq;
-using BH.oM.Geometry;
 using BH.oM.Structure.Elements;
-using BH.oM.Structure.Loads;
 using Lusas.LPI;
 
 namespace BH.Engine.Lusas
 {
-    public static partial class Convert
+    public partial class Query
     {
-        public static PointLoad ToPointLoad(
-            IFLoading lusasPointLoad, IEnumerable<IFAssignment> lusasAssignments, 
-            Dictionary<string, Node> bhomNodeDictionary)
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
+
+        public static Edge GetEdge(IFSurface lusasSurf, int lineIndex, Dictionary<string, Edge> bhomBars)
         {
-            IFLoadcase assignedLoadcase = (IFLoadcase)lusasAssignments.First().getAssignmentLoadset();
-            Loadcase bhomLoadcase = ToLoadcase(assignedLoadcase);
-
-            IEnumerable<Node> bhomNodes = Lusas.Query.GetPointAssignments(lusasAssignments, bhomNodeDictionary);
-
-            Vector forceVector = new Vector
-            {
-                X = lusasPointLoad.getValue("px"),
-                Y = lusasPointLoad.getValue("py"),
-                Z = lusasPointLoad.getValue("pz")
-            };
-
-            Vector momentVector = new Vector
-            {
-                X = lusasPointLoad.getValue("mx"),
-                Y = lusasPointLoad.getValue("my"),
-                Z = lusasPointLoad.getValue("mz")
-            };
-
-            PointLoad bhomPointLoad = Structure.Create.PointLoad(
-                bhomLoadcase,
-                bhomNodes,
-                forceVector,
-                momentVector,
-                LoadAxis.Global,
-                Lusas.Query.GetName(lusasPointLoad));
-
-            int adapterID = Lusas.Query.GetAdapterID(lusasPointLoad, 'l');
-            bhomPointLoad.CustomData["Lusas_id"] = adapterID;
-
-            return bhomPointLoad;
+            Edge bhomEdge = null;
+            IFLine lusasEdge = lusasSurf.getLOFs()[lineIndex];
+            string lineName = Engine.Lusas.Modify.RemovePrefix(lusasEdge.getName(), "L");
+            bhomBars.TryGetValue(lineName, out bhomEdge);
+            return bhomEdge;
         }
+
+        /***************************************************/
+
     }
 }
