+++ Lusas_Engine/Query/GetGeometryAssignment.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -29,78 +29,9 @@ namespace BH.Engine.Lusas
 {
     public partial class Query
     {
-        public static IEnumerable<Node> GetNodeAssignments(IEnumerable<IFAssignment> lusasAssignments,
-    Dictionary<string, Node> bhomNodes)
-        {
-            List<Node> assignedNodes = new List<Node>();
-            Node bhomNode = new Node();
-
-            foreach (IFAssignment lusasAssignment in lusasAssignments)
-            {
-                if (lusasAssignment.getDatabaseObject() is IFPoint)
-                {
-                    IFPoint lusasPoint = (IFPoint)lusasAssignment.getDatabaseObject();
-                    bhomNodes.TryGetValue(Engine.Lusas.Modify.RemovePrefix(lusasPoint.getName(), "P"), out bhomNode);
-                    assignedNodes.Add(bhomNode);
-                }
-                else
-                {
-                    Compute.WarningLineAssignment(lusasAssignment);
-                    Compute.WarningSurfaceAssignment(lusasAssignment);
-                }
-            }
-
-            return assignedNodes;
-        }
-
-        public static IEnumerable<Bar> GetBarAssignments(IEnumerable<IFAssignment> lusasAssignments,
-            Dictionary<string, Bar> bhomBars)
-        {
-            List<Bar> assignedBars = new List<Bar>();
-            Bar bhomBar = new Bar();
-
-            foreach (IFAssignment lusasAssignment in lusasAssignments)
-            {
-                if (lusasAssignment.getDatabaseObject() is IFLine)
-                {
-                    IFLine lusasLine = (IFLine)lusasAssignment.getDatabaseObject();
-                    bhomBars.TryGetValue(Engine.Lusas.Modify.RemovePrefix(lusasLine.getName(), "L"), out bhomBar);
-                    assignedBars.Add(bhomBar);
-                }
-                else
-                {
-                    Lusas.Compute.WarningPointAssignment(lusasAssignment);
-                    Lusas.Compute.WarningSurfaceAssignment(lusasAssignment);
-                }
-
-            }
-
-            return assignedBars;
-        }
-
-        public static IEnumerable<IAreaElement> GetSurfaceAssignments(IEnumerable<IFAssignment> lusasAssignments,
-            Dictionary<string, Panel> bhomPanels)
-        {
-            List<IAreaElement> assignedSurfs = new List<IAreaElement>();
-            Panel bhomPanel = new Panel();
-
-            foreach (IFAssignment lusasAssignment in lusasAssignments)
-            {
-                if (lusasAssignment.getDatabaseObject() is IFSurface)
-                {
-                    IFSurface lusasSurface = (IFSurface)lusasAssignment.getDatabaseObject();
-                    bhomPanels.TryGetValue(Engine.Lusas.Modify.RemovePrefix(lusasSurface.getName(), "S"), out bhomPanel);
-                    assignedSurfs.Add(bhomPanel);
-                }
-                else
-                {
-                    Lusas.Compute.WarningPointAssignment(lusasAssignment);
-                    Lusas.Compute.WarningLineAssignment(lusasAssignment);
-                }
-            }
-
-            return assignedSurfs;
-        }
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
 
         public static IEnumerable<BHoMObject> GetGeometryAssignments(IEnumerable<IFAssignment> lusasAssignments,
             Dictionary<string, Node> bhomNodes, Dictionary<string, Bar> bhomBars,
@@ -135,5 +66,8 @@ namespace BH.Engine.Lusas
 
             return assignedObjects;
         }
+
+        /***************************************************/
+
     }
 }
